Solve TroubleSort
The idea is that trouble sort is essentially a bubble sort on two separate
lists - one with even numbered indices and one with odd. If we separate those
two lists and efficiently sort them (using the built in sort) then we could
rejoin the two lists and simply check if there is an issue and where. This
solution doesn't really combine it back, however, but simply traverses both
lists together and finds the faulty index.