Add another test case
This one shows where my current idea of the final solution would go wrong. It
would select a sub-optimal solution, because the first bit for the second
cashier is more expensive than the first bit for any others and so it won't be
considered, although it is required for the correct solution, because at least
one cashier needs to process 2 bits and other cashiers process it with time 10,
but the second cashier would process it with 8.