Solve Edgy Baking for Test set 1 (Visible)
This algorithm only works if all cookies are of the same size. If the cookie
sizes vary, then this might not find the highest possible value, because it
tries adding the cookies in the given order and gives up once it reaches or
surpasses the goal. A different order might give a better result.