Solve Edgy Baking
The idea is, that the list of possible ranges a list of cookies can be cut into
stays quite small, because maximum width and height are quite small. This means
that the combinations will overlap and can be merged and so the list of ranges
doesn't actually grow combinatorially. Knowing this, we can calculate the
sorted list of possible ranges of perimeter values the cookies can cut into and
then traverse that list to find a maximal value that is not above the goal.

Took me a while and I considered giving up and reading the posted problem
analysis from the competition site a couple of times, but finally figured it
out.